refactor: make pipeline modules crate-internal; inline parser test
